# Go_Server/docs/swagger.yaml
definitions:
...
  define.AddUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      remarks:
        type: string
      role_id:
        type: integer
      sex:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
...
  define.GetUserListRequest:
    properties:
      keyword:
        type: string
      page:
        type: integer
      size:
        type: integer
      status:
        type: integer
    type: object
...
  define.UpdateUserRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
      remarks:
        type: string
      role_id:
        type: integer
      sex:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
info:
  contact:
    name: Vulnerabilities_Server
    url: https://github.com/A7cc/Vulnerabilities_Server
  description: 这是一个用Golang写的食谱菜单管理系统靶场，该系统中存在多个安全漏洞，需要我们去探索和发现。\n\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于Golang系统的渗透测试和代码审计知识。
  title: 食谱菜单管理系统靶场
  version: "1.0"
paths:
...
  /user/add:
    post:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 添加用户信息
        in: body
        name: AddUserRequest
        required: true
        schema:
          $ref: '#/definitions/define.AddUserRequest'
      responses: {}
      summary: 新增用户信息
      tags:
      - 鉴权接口-用户相关方法
  /user/delete/{id}:
    delete:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除用户ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除用户信息
      tags:
      - 鉴权接口-用户相关方法
  /user/detail:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取用户ID
        in: query
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据ID获取用户信息
      tags:
      - 鉴权接口-用户相关方法
  /user/get:
    get:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取用户查询参数
        in: body
        name: GetUserListRequest
        required: true
        schema:
          $ref: '#/definitions/define.GetUserListRequest'
      responses: {}
      summary: 获取用户列表信息
      tags:
      - 鉴权接口-用户相关方法
  /user/update:
    put:
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新用户信息参数
        in: body
        name: UpdateUserRequest
        required: true
        schema:
          $ref: '#/definitions/define.UpdateUserRequest'
      responses: {}
      summary: 修改用户信息
      tags:
      - 鉴权接口-用户相关方法
swagger: "2.0"